+++ Extracting_keywords/GUI/AlgorithmsEstimationForm.py
@@ -51,3 +51,4 @@ class AlgorithmsEstimationForm(QWidget):
         # self.setGeometry(300, 300, 400, 200)
         # self.setWindowTitle('Оценка работы алгоритма')
         # self.show()
+
